Closes guard/guard-minitest#84: Shelling out with clean env when not using bundler
Even with the `bundler` option set to false (or negated by the `zeus` or
`spring` options), if Bundler has been loaded prior to shelling out, it
somehow alters/infects the new environment. The solution is to wrap
the logic in a `Bundler.with_clean_env` block.

Also notifying guard with success/fail for spring as well as zeus.