Add env setting options for all runs and full runs
Add two options:
  * env
  * all_env

`env` is a Hash which adds to the environment for every run.
This can be used, for example, to generally indicate to your tests
that they are being run in the context of Guard if this is helpful.

`all_env` is merged into the environment when running the full suite.
This can be used, for example, to signal that coverage can be run.