Fix plugin auto-discovery by adding a --guard flag
Minitest loads all plugins regardless of context and this caused guard
notifiers to fire even when tests were run outside of guard. To fix this
we pass a --guard flag to minitest and only add the
Guard::Minitest::Reporter if the flag is set.

Also the loading of guard related code is deferred and controlled by the
--guard flag. This way guard and everything that comes with it will not
be loaded outside of guard.